requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
mysql-connector-python==8.0.26
pytesseract==0.3.8
opencv-python==4.5.3.56
numpy==1.21.2
pdf2image==1.16.0
PyPDF2==2.0.0
Pillow==8.3.2
gunicorn==20.1.0
python-dotenv==0.19.0 